+++ Project.toml
@@ -11,6 +11,7 @@ AxisIndices = "f52c9ee2-1b1c-4fd8-8546-6350938c7f11"
 Dictionaries = "85a47980-9c8c-11e8-2b9f-f7ca1fa99fb4"
 Distances = "b4f34e82-e78d-54a5-968a-f98e89d6e8f7"
 EcoBase = "a58aae7d-b440-5a11-b283-399458f99aac"
+InlineTest = "bd334432-b1e7-49c7-a2dc-dd9149e4ebd6"
 MultivariateStats = "6f286f6a-111f-5878-ab1e-185364afe411"
 NamedDims = "356022a1-0364-5f58-8944-0da4b18d706f"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
